feat(ci): update deployment stages and Docker scripts
Removed the `deploy-staging` stage and added `deploy-sec`
to run only on the `main` branch. Adjusted `before_script`
and `after_script` in `docker-dev` for consistent Docker
login and image removal. Added dependency on `docker-dev`
in `deploy-new-prd` to ensure images are built and pushed
before deployment. Updated scripts in `docker-prd` for
consistency.