feat(OperationLog): Improve operator name handling (#1442)
* feat(OperationLog): Improve operator name handling

Updated the logic for retrieving the operator's name in
`OperationLogRepository.cs` to use the operator's string
representation if the real display name is null. Also ensured
the `clientId` parameter is passed when calling `AddAsync`.

Modified the `RegisterUserOperationLogAsync` method in
`OperationLogCommandHandler.cs` to include the `clientId`
parameter when logging user registration operations.

* feat(OperationLogRepository): improve operatorName assignment

Updated the assignment logic for `operatorName` in
`OperationLogRepository.cs` to use `@operator.Value.ToString()`
instead of `@operator.ToString()`. This change ensures that
`operatorName` is more accurate when `@operator` is
`Guid.Empty`. Additionally, removed the null-forgiving operator
from the `AddAsync` method call to enhance code safety and
readability.